Fixed the abusive usage of GST_STATE_NULL to pause all cached sounds when a new sound play request is made.
According to this document:
https://gstreamer.freedesktop.org/documentation/plugin-development/basics/states.html
The GST_STATE_NULL should be used olny when no ressources is loaded/loaded.
The GST_STATE_PAUSED fits best to the requirement to have olny one sound
played at time.

Concerning the bug of the jumping sound volume, it has been fixed by replacing
the usage of GST_STATE_NULL by GST_STATE_PAUSED.

Concerning the whole usage of the GST library this commit is just a "hot fix"
I think the GST library should not be used the way it is now, in my opinion
as we want to have olny one sound playing a time we need a unique sound
stream/channel and a unique gstsoud buffer containing multiple sounds.
In this buffer we should keep track off the sounds addresses. By this way to
play a given sound we just have to seek in the sound buffer.